Moved SubCommands into separate class files
